fix type error in pipeline index change

+++ photon-core/src/main/java/org/photonvision/common/dataflow/networktables/NTDataPublisher.java
@@ -91,7 +91,7 @@ public class NTDataPublisher implements CVPipelineResultConsumer {
     }
 
     private void onPipelineIndexChange(NetworkTableEvent entryNotification) {
-        var newIndex = (int) entryNotification.valueData.value.getDouble();
+        var newIndex = (int) entryNotification.valueData.value.getInteger();
         var originalIndex = pipelineIndexSupplier.get();
 
         // ignore indexes below 0
